% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{apartment}
\alias{apartment}
\title{中古マンションデータ}
\format{
5390行 4列のデータフレーム:
\describe{
\item{price}{取引価格 (100万円)}
\item{tenure}{築年数 (年)}
\item{number}{部屋数}
\item{reform}{改築状況 (=1: 改築済み, =0: 未改築)}
}
}
\source{
\url{https://example.com/data}
}
\usage{
apartment
}
\description{
このデータセットは計量経済学basicの実習用データです
}
\keyword{datasets}
